// src/api/generated/models/workflowControllerFindAllSortOrder.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Flowforge API
 * The Flowforge workflow orchestration engine API documentation
 * OpenAPI spec version: 1.0
 */

export type WorkflowControllerFindAllSortOrder = typeof WorkflowControllerFindAllSortOrder[keyof typeof WorkflowControllerFindAllSortOrder];


export const WorkflowControllerFindAllSortOrder = {
  asc: 'asc',
  desc: 'desc',
} as const;