// src/api/generated/models/workflowListResponseDto.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Flowforge API
 * The Flowforge workflow orchestration engine API documentation
 * OpenAPI spec version: 1.0
 */
import type { PaginationMetaDto } from './paginationMetaDto';
import type { WorkflowDataDto } from './workflowDataDto';

export interface WorkflowListResponseDto {
  message: string;
  data: WorkflowDataDto[];
  meta: PaginationMetaDto;
}